// dashboard/src/api/models/invoiceStatusEnum.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Invoicence API
 * Description
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `draft` - Draft
 * `open` - Open
 * `paid` - Paid
 * `voided` - Voided
 */
export type InvoiceStatusEnum =
  (typeof InvoiceStatusEnum)[keyof typeof InvoiceStatusEnum];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const InvoiceStatusEnum = {
  draft: "draft",
  open: "open",
  paid: "paid",
  voided: "voided",
} as const;